+++ gazebo/gui/building/EditorItem.hh
@@ -226,8 +226,12 @@ namespace gazebo
       /// \brief Transparency of the associated 3D visual.
       protected: float visual3dTransparency;
 
-      /// \brief TODO
+      /// \brief Flag to indicate whether this item is currently highlighted or
+      /// not.
       protected: bool highlighted;
+
+      /// \brief Scale for converting from pixel to metric units.
+      protected: double scale;
     };
     /// \}
   }