+++ gazebo/gui/building/RectItem.cc
@@ -33,15 +33,12 @@ RectItem::RectItem()
   this->width = 100;
   this->height = 100;
   this->highlighted = true;
-  this->visual3dColor = QColor(255, 255, 255, 255);
-  this->visual3dTransparency = 0.0;
 
   this->drawingOriginX = 0;
   this->drawingOriginY = 0;
 
   this->positionOnWall = 0;
   this->angleOnWall = 0;
-  this->parentWall = NULL;
 
   this->drawingWidth = this->width;
   this->drawingHeight = this->height;
@@ -160,13 +157,8 @@ void RectItem::SetHighlighted(bool _highlighted)
     this->rotateHandle->removeSceneEventFilter(this);
     this->Set3dTransparency(0.4);
   }
-  for (unsigned int j = 0; j < this->measures.size(); ++j)
-  {
-    this->measures[j]->setVisible(_highlighted && (this->parentWall != NULL));
-  }
   this->highlighted = _highlighted;
-  emit TransparencyChanged(this->visual3dTransparency);
-  this->RectUpdated();
+  this->UpdateMeasures();
 }
 
 /////////////////////////////////////////////////
@@ -492,10 +484,23 @@ bool RectItem::GrabberEventFilter(GrabberHandle *_grabber, QEvent *_event)
         dx = cos(-angle) * deltaWidth/2;
         dy = -sin(-angle) * deltaWidth/2;
         this->SetPosition(this->pos() - QPointF(dx, dy));
-        if (this->parentWall)
+        if (this->parentItem())
         {
-          this->positionOnWall -= deltaWidth /
-              (2*this->parentWall->line().length());
+          WallSegmentItem *wallItem = dynamic_cast<WallSegmentItem *>(
+              this->parentItem());
+          if (wallItem)
+          {
+            if (this->GetAngleOnWall() < 90)
+            {
+              this->positionOnWall -= deltaWidth /
+                  (2*wallItem->line().length());
+            }
+            else
+            {
+              this->positionOnWall += deltaWidth /
+                  (2*wallItem->line().length());
+            }
+          }
         }
         break;
       }
@@ -511,10 +516,23 @@ bool RectItem::GrabberEventFilter(GrabberHandle *_grabber, QEvent *_event)
         dx = cos(angle) * deltaWidth/2;
         dy = sin(angle) * deltaWidth/2;
         this->SetPosition(this->pos() + QPointF(dx, dy));
-        if (this->parentWall)
+        if (this->parentItem())
         {
-          this->positionOnWall += deltaWidth /
-              (2*this->parentWall->line().length());
+          WallSegmentItem *wallItem = dynamic_cast<WallSegmentItem *>(
+              this->parentItem());
+          if (wallItem)
+          {
+            if (this->GetAngleOnWall() < 90)
+            {
+              this->positionOnWall += deltaWidth /
+                  (2*wallItem->line().length());
+            }
+            else
+            {
+              this->positionOnWall -= deltaWidth /
+                  (2*wallItem->line().length());
+            }
+          }
         }
         break;
       }
@@ -720,12 +738,10 @@ double RectItem::GetHeight() const
 }
 
 /////////////////////////////////////////////////
-void RectItem::SetPositionOnWall(double _positionOnWall,
-    WallSegmentItem *_wall)
+void RectItem::SetPositionOnWall(double _positionOnWall)
 {
   this->positionOnWall = _positionOnWall;
-  this->parentWall = _wall;
-  this->RectUpdated();
+  this->UpdateMeasures();
 }
 
 /////////////////////////////////////////////////
@@ -738,7 +754,7 @@ double RectItem::GetPositionOnWall() const
 void RectItem::SetAngleOnWall(double _angleOnWall)
 {
   this->angleOnWall = _angleOnWall;
-  this->RectUpdated();
+  this->UpdateMeasures();
 }
 
 /////////////////////////////////////////////////
@@ -878,7 +894,7 @@ void RectItem::SizeChanged()
 {
   emit DepthChanged(this->drawingHeight);
   emit WidthChanged(this->drawingWidth);
-  this->RectUpdated();
+  this->UpdateMeasures();
 }
 
 /////////////////////////////////////////////////
@@ -912,7 +928,56 @@ void RectItem::SetResizeFlag(unsigned int _flag)
 }
 
 /////////////////////////////////////////////////
-void RectItem::RectUpdated()
+void RectItem::UpdateMeasures()
 {
-  // virtual
+  // Only windows and doors can have a wall as parent
+  WallSegmentItem *wallItem = dynamic_cast<WallSegmentItem *>(
+    this->parentItem());
+  if (wallItem == NULL)
+  {
+    for (unsigned int i = 0; i < this->measures.size(); ++i)
+    {
+      this->measures[i]->setVisible(false);
+    }
+    return;
+  }
+
+  if (this->measures.empty())
+  {
+    this->measures.push_back(new MeasureItem(QPointF(0, 0), QPointF(0, 1)));
+    this->measures.push_back(new MeasureItem(QPointF(0, 0), QPointF(0, 1)));
+    this->measures[0]->setParentItem(this);
+    this->measures[1]->setParentItem(this);
+    this->measures[0]->setVisible(false);
+    this->measures[1]->setVisible(false);
+  }
+
+  this->measures[0]->setVisible(this->highlighted);
+  this->measures[1]->setVisible(this->highlighted);
+
+  if (this->highlighted)
+  {
+    double d = 20 + this->drawingHeight/2.0;
+    double t = wallItem->GetThickness()/2.0;
+    double l = wallItem->line().length();
+    double p = this->GetPositionOnWall();
+
+    if (this->GetAngleOnWall() < 90)
+    {
+      this->measures[0]->SetStartPoint(QPointF(-(t + l*p), -d));
+      this->measures[1]->SetEndPoint(QPointF(t + l*(1-p), -d));
+    }
+    else
+    {
+      this->measures[0]->SetStartPoint(QPointF(-(t + l*(1-p)), -d));
+      this->measures[1]->SetEndPoint(QPointF(t + l*p, -d));
+    }
+    this->measures[0]->SetEndPoint(QPointF(-this->drawingWidth/2, -d));
+    this->measures[1]->SetStartPoint(QPointF(this->drawingWidth/2, -d));
+
+    this->measures[0]->SetValue(
+        (this->measures[0]->line().length())*this->scale);
+    this->measures[1]->SetValue(
+        (this->measures[1]->line().length())*this->scale);
+  }
 }