+++ gazebo/gui/building/RectItem.hh
@@ -73,20 +73,17 @@ namespace gazebo
 
       /// \brief Set the position of this item inside its parent wall.
       /// \param[in] _positionOnWall New normalized position on wall.
-      /// TODO
-      public: void SetPositionOnWall(double _positionOnWall,
-          WallSegmentItem *_wall);
+      public: void SetPositionOnWall(double _positionOnWall);
 
       /// \brief Get the position of this item inside its parent wall.
       /// \return Normalized position on parent wall.
       public: double GetPositionOnWall() const;
 
-      /// \brief Set the amgle of this item inside its parent wall.
-      /// \param[in] _engleOnWall New angle on wall, either 0 or 180 degrees.
-      /// TODO
+      /// \brief Set the angle of this item inside its parent wall.
+      /// \param[in] _angleOnWall New angle on wall, either 0 or 180 degrees.
       public: void SetAngleOnWall(double _angleOnWall);
 
-      /// \brief Get the amgle of this item on its parent wall.
+      /// \brief Get the angle of this item inside its parent wall.
       /// \return Angle on parent wall in degrees.
       public: double GetAngleOnWall() const;
 
@@ -140,8 +137,8 @@ namespace gazebo
       /// \return The bounding box of the rect item.
       protected: virtual QRectF boundingRect() const;
 
-      /// \brief Update item.
-      protected: virtual void RectUpdated();
+      /// \brief Update this item's measures.
+      protected: void UpdateMeasures();
 
       /// \brief Filter Qt events and redirect them to the rotate handle.
       /// \param[in] _rotateHandle Rotate handle that will handle the event.
@@ -267,11 +264,13 @@ namespace gazebo
       /// \brief Qt action for deleting the item.
       protected: QAction *deleteItemAct;
 
-      /// \brief TODO
-      protected: WallSegmentItem *parentWall;
-
-      /// \brief TODO
+      /// \brief A vector containing this item's measure items.
+      /// Currently only used for windows and doors, containing one measure
+      /// towards each end of this item's parent wall.
       protected: std::vector<MeasureItem *> measures;
+
+      /// \brief Scale for converting from pixel to metric units.
+      protected: double scale;
 
       /// \brief Mouse press position in pixel coordinates.
       private: QPointF mousePressPos;
@@ -300,7 +299,7 @@ namespace gazebo
       /// point.
       private: double positionOnWall;
 
-      /// \brief TODO
+      /// \brief Angle with respect to parent wall, either 0 or 180 degrees.
       private: double angleOnWall;
     };
     /// \}