+++ gazebo/gui/building/WindowItem.hh
@@ -96,9 +96,6 @@ namespace gazebo
         /// \brief Elevation of the window item in pixels.
         private: double windowElevation;
 
-        /// \brief Scale for converting pixels to metric units.
-        private: double scale;
-
         /// \brief Inspector for configuring the window item.
         private: WindowDoorInspectorDialog *inspector;
     };